Refactor MicroBlockProcessing into Pre and Post stages
